Finished Delete Functions. Not including sessions

+++ functions/delete-account.php
@@ -1,16 +1,23 @@
 <?php
 
-require_once '../includes.php';
-require_once '../database.php';
+if (isset($_POST['delete-account-modal__btn-delete-confirm'])) {
+  require_once '../includes.php';
+  require_once '../database.php';
 
-$accountId = $_POST['id'];
+  $accountId = 2;
+  $studentId = 1;
 
-$accountClass = new AccountDAO();
-$deleteAccount = $accountClass->deleteAccount($pdo, $accountId);
+  $accountClass = new AccountDAO();
+  $deleteAccount = $accountClass->deleteAccount($pdo, $accountId);
+  $deleteEnrollement = $accountClass->deleteEnrollement($pdo, $studentId);
+  $deleteStudent = $accountClass->deleteStudent($pdo, $studentId);
+  $deleteStudentStack = $accountClass->deleteStudentStack($pdo, $studentId);
+  $deleteStudentTech = $accountClass->deleteStudentTech($pdo, $studentId);
+  $deleteStudentExternals = $accountClass->deleteStudentExternals($pdo, $studentId);
+  $deleteStudentProjects = $accountClass->deleteStudentProjects($pdo, $studentId);
 
-$jDeleteProject = json_encode($deleteAccount);
-
-header("Content-Type: application/json");
-echo $jDeleteProject;
+  header('location: ../account-settings.php');
+  
+}
 
 ?>
